requirements.txt
Flask==2.3.2
tensorflow==2.12.0
numpy
Pillow
